+++ ToolBox/Log/ILogSystem.cs
@@ -3,7 +3,7 @@ using System;
 namespace ToolBox.Log
 {
     public interface ILogSystem
-    
+
     {
         void AccessValidation();
         void Save(Exception ex, LogLevel logLevel = LogLevel.Information);