+++ ToolBox/System/Env.cs
@@ -4,15 +4,18 @@ namespace ToolBox.System
 {
     public static class Env
     {
-        public static string GetValue(string name) {
+        public static string GetValue(string name)
+        {
             return Environment.GetEnvironmentVariable(name);
         }
 
-        public static void SetValue(string name, string value) {
+        public static void SetValue(string name, string value)
+        {
             Environment.SetEnvironmentVariable(name, value);
         }
 
-        public static bool IsNullOrEmpty(string name){
+        public static bool IsNullOrEmpty(string name)
+        {
             string env = Env.GetValue(name);
             return String.IsNullOrEmpty(env);
         }