+++ ToolBox/System/User.cs
@@ -5,15 +5,18 @@ namespace ToolBox.System
     public static class User
     {
 
-        public static string GetUserName(){
+        public static string GetUserName()
+        {
             return Env.GetValue("USERNAME") ?? Env.GetValue("USER");
         }
 
-        public static string GetMachine(){
+        public static string GetMachine()
+        {
             return Environment.MachineName;
         }
 
-        public static string GetDomain(){
+        public static string GetDomain()
+        {
             return Env.GetValue("USERDOMAIN") ?? Env.GetValue("HOSTNAME");
         }
     }